# apps/docs/api-reference/openapi.yaml
openapi: 3.1.0
info:
  title: Authlane Control Plane API
  version: 1.0.0
  description: Manages SaaS tenant policy, end-user connections, credentials, and tool definitions. Provider tool traffic does not pass through this API.
servers:
  - url: https://authlane.example.com
  - url: http://localhost:3000
security:
  - apiKey: []
paths:
...
  /api/v1/connect-sessions:
    post:
      operationId: createConnectSession
      summary: Create a short-lived browser handoff
      x-authlane-scope: connect-sessions:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [externalUserId, allowedServices, allowedOrigin]
              properties:
                externalUserId:
                  type: string
                allowedServices:
                  type: array
                  minItems: 1
                  items:
                    type: string
                allowedOrigin:
                  type: string
                  format: uri
                expiresInSeconds:
                  type: integer
                  minimum: 60
                  maximum: 900
                  default: 600
      responses:
        '201':
          description: Raw token is returned once for browser handoff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectSessionResult'
  /api/v1/connect/{serviceId}/authorize:
    post:
      security: []
      operationId: authorizeConnection
      summary: Start OAuth from a connect session
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectAction'
      responses:
        '200':
          description: Provider authorization URL
    delete:
      security: []
      operationId: disconnectService
      summary: Disconnect through a connect session
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectAction'
      responses:
        '200':
          description: Disconnect result
  /api/v1/oauth/{serviceId}/callback:
    get:
      security: []
      operationId: oauthCallback
      summary: Validate state and exchange the provider code
      parameters:
        - $ref: '#/components/parameters/ServiceId'
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the hosted connect callback
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: Authlane scoped API key
  parameters:
...
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9-]+$'
...
  schemas:
...
    ConnectAction:
      type: object
      required: [connectToken, parentOrigin]
      properties:
        connectToken:
          type: string
        parentOrigin:
          type: string
          format: uri
    ConnectSessionResult:
      type: object
      required: [data, error]
      properties:
        data:
          type: object
          required: [id, token, url, expiresAt]
          properties:
            id:
              type: string
            token:
              type: string
            url:
              type: string
              format: uri
            expiresAt:
              type: string
              format: date-time
        error:
          type: 'null'
...